+++ src/Console/origin.php
@@ -17,7 +17,7 @@ use Origin\Console\ShellDispatcher;
 use Origin\Console\ConsoleOutput;
 use Origin\Console\ConsoleInput;
 
-require dirname(__DIR__) . '/bootstrap.php';
+require dirname(__DIR__) . DIRECTORY_SEPARATOR . 'bootstrap.php';
 
 $Dispatcher = new ShellDispatcher($argv, new ConsoleOutput(), new ConsoleInput());
 $Dispatcher->start();